Refactor imports to use main_entry and enhance HTML rendering logic with headless browser integration
