+++ .vscode/settings.json
@@ -1,5 +1,11 @@
 {
     "github-actions.workflows.pinned.workflows": [
         ".github/workflows/CI.yml"
+    ],
+    "python.analysis.extraPaths": [
+        "/home/codespace/.vscode-remote/extensions/continue.continue-0.1.3"
+    ],
+    "python.autoComplete.extraPaths": [
+        "/home/codespace/.vscode-remote/extensions/continue.continue-0.1.3"
     ]
 }
